requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-decouple==3.8
psycopg2-binary==2.9.7
drf-spectacular==0.27.0
# For production PostgreSQL support
# For demo/development uses built-in SQLite